# stdaddr: Introduce package for standard addrs.