--- a/internal/staging/stdaddr/address.go
+++ b/internal/staging/stdaddr/address.go
@@ -123,9 +123,16 @@ type Secp256k1PublicKey interface {
 //
 // This function can be useful to callers who already need the serialized public
 // key for other purposes to avoid the need to serialize it multiple times.
+//
+// NOTE: Version 0 scripts are the only currently supported version.
 func NewAddressPubKeyEcdsaSecp256k1Raw(scriptVersion uint16,
 	serializedPubKey []byte,
 	params AddressParams) (Address, error) {
+
+	switch scriptVersion {
+	case 0:
+		return NewAddressPubKeyEcdsaSecp256k1V0Raw(serializedPubKey, params)
+	}
 
 	str := fmt.Sprintf("pubkey addresses for version %d are not supported",
 		scriptVersion)
@@ -140,9 +147,16 @@ func NewAddressPubKeyEcdsaSecp256k1Raw(scriptVersion uint16,
 // key already serialized in the _compressed_ format instead of a concrete type.
 // It can be useful to callers who already need the serialized public key for
 // other purposes to avoid the need to serialize it multiple times.
+//
+// NOTE: Version 0 scripts are the only currently supported version.
 func NewAddressPubKeyEcdsaSecp256k1(scriptVersion uint16,
 	pubKey Secp256k1PublicKey,
 	params AddressParams) (Address, error) {
+
+	switch scriptVersion {
+	case 0:
+		return NewAddressPubKeyEcdsaSecp256k1V0(pubKey, params)
+	}
 
 	str := fmt.Sprintf("pubkey addresses for version %d are not supported",
 		scriptVersion)