The current code for handling standard addresses implemented in `dcrutil` was written many years ago prior a wide variety of changes and several new features added by Decred.  As a result, it entirely lacks support for some features and supports others in a roundabout and non-intuitive way.

Specifically, it does not support or provide a clean path to enable support for different script versions and the way they are handled in stake transactions is entirely non-intuitive.

Further, back when the original address code was implemented, it was necessary to implement script creation in the `txscript` package which led to the current design of providing methods such as `txscript.PayTo{AddrScript,SStx,SStxChange}`, and others, which need to type assert the specific concrete types of addresses in order to produce the necessary scripts.  This, unfortunately, effectively negates the use of an interface to support generic addresses because it means callers, such as `dcrwallet`, are not able to implement their own types without somewhat invisibly breaking the script creation.

Finally, the aforementioned blending of the address code into `txscript` has led to confusion regarding what is considered standard and what is considered consensus which has tripped up several contributors over the years.

This aims to resolve all of the aforementioned issues by introducing a new package named `stdaddr` which entirely reworks the way addresses are handled.

It is split up into a series of commits to help ease the review process such that the initial generic infrastructure is introduced first, followed by the overall version 0 address decoding logic, then each supported version 0 address, followed by an example, and finally what I hope is a fairly complete `README.md` that describes the architecture.  Comprehensive tests are included.

For the time being, the package is introduced into the internal staging area for initial review.

The following provides an overview of some of the key features of the new design:

- Supports versioned addresses
- Produces scripts directly via methods implemented on underlying types
- Provides direct support for creation of the scripts necessary for the staking system
- Uses a capabilities-based approach via interfaces so callers can cleanly and generically determine under what circumstances addresses can be used
- Allows callers to create their own concrete address types without worrying about breaking the existing ones
- Clearly denotes that addresses are a standardized construction that must not be used directly in consensus code

Finally, although the primary goal here is not optimization, this also significantly reduces the execution time and number of allocations needed to produce the payment scripts as can be seen in the following benchmarks:

Name                       | old time/op | new time/op | delta
---------------------------|-------------|-------------|------
p2sh                       | 207ns ± 1%  | 39ns ± 1%   | -81.24%  (p=0.008 n=5+5)
p2pkh-ecdsa-secp256k1      | 219ns ± 2%  | 45ns ± 8%   | -79.44%  (p=0.008 n=5+5)
p2pkh-ed25519              | 235ns ± 2%  | 42ns ± 4%   | -82.30%  (p=0.008 n=5+5)
p2pkh-schnorr-secp256k1    | 232ns ± 1%  | 42ns ± 3%   | -81.99%  (p=0.008 n=5+5)
p2pk-ecdsa-secp256k1       | 276ns ± 1%  | 47ns ± 2%   | -83.10%  (p=0.008 n=5+5)
p2pk-ed25519               | 665ns ± 6%  | 47ns ± 1%   | -92.98%  (p=0.008 n=5+5)
p2pk-schnorr-secp256k1     | 283ns ± 1%  | 47ns ± 4%   | -83.47%  (p=0.008 n=5+5)

Name                    | old alloc/op | new alloc/op | delta
------------------------|--------------|--------------|------
p2sh                    | 512B ± 0%    | 24B ± 0%     | -95.31%  (p=0.008 n=5+5)
p2pkh-ecdsa-secp256k1   | 512B ± 0%    | 32B ± 0%     | -93.75%  (p=0.008 n=5+5)
p2pkh-ed25519           | 512B ± 0%    | 32B ± 0%     | -93.75%  (p=0.008 n=5+5)
p2pkh-schnorr-secp256k1 | 512B ± 0%    | 32B ± 0%     | -93.75%  (p=0.008 n=5+5)
p2pk-ecdsa-secp256k1    | 560B ± 0%    | 48B ± 0%     | -91.43%  (p=0.008 n=5+5)
p2pk-ed25519            | 640B ± 0%    | 48B ± 0%     | -92.50%  (p=0.008 n=5+5)
p2pk-schnorr-secp256k1  | 560B ± 0%    | 48B ± 0%     | -91.43%  (p=0.008 n=5+5)