@jrick 2021-03-05
This looks great.

May i suggest naming the package `stdaddr`? A little shorter and unlikely to collide with any variable names because those would probably read `stdAddr` when following the usual camel case style.