@davecgh 2021-03-05
> 
> 
> This looks great.
> 
> May i suggest naming the package `stdaddr`? A little shorter and unlikely to collide with any variable names because those would probably read `stdAddr` when following the usual camel case style.

That sounds good to me.  We also need to decide on where to put it.  I just put it under `internal/staging` for now, but it will need to be exposed.