@davecgh 2021-03-06
I've updated this to rename the package to the recommended `stdaddr` and updated all references in the code, commits, and documentation accordingly.