@davecgh 2021-03-06
@matthawkins90 Thanks for the feedback.  It's always good to get a fresh pair of eyes!

I'll comment inline, as your points are all valid, but, I did want to point out that, as you note, this is really aimed at developers who already have some background knowledge about the underlying concepts.  If this were in a developer guide, I would put this type of information after sections which have already describing the how the transaction scripting language works which includes opcodes.

> You don't really define `script` in the Readme...

I'll add a link to the `txscript` package which should help clarify this I think.

> Is Decred's script the same as Bitcoin's, just with different opcodes?

Yes.

> You don't really talk about opcodes until `Address Use in the Staking System`

Right.  The underlying transaction scripts are really just a collection of opcodes.  There is some assumed knowledge here.  Again, I think having an initial link to the `txscript` package near the top will help with this since it provides the prerequisite knowledge.

> You use `script` instead of `Script`. The capitalization seems like it might be more clear that you're referring to the Script system.

Not sure I follow you here.  Is there a specific instance you can point out?  As far as I can tell, all of the references to scripts are indeed talking about scripts, whereas references to the scripting system specifically say scripting system.  Maybe I missed some cases?

> 
> I don't know what Hash160 is, even after reading the blurb about RIPEMD-160 hashes. It would be nice to link to a reference.

Good point.  I'll add a link to the `ripemd160` package and clarify that `Hash160` is shorthand for `RIPEMD-160(BLAKE256(data))`.

> I don't know what you mean by Version 0 addresses. I don't know much about address versioning, unless you're talking about things like Segwit or Taproot.

That is likely because there  currently is no real notion of versioned addresses since there is only the existing addresses (hence version 0).  The plan in the future is to introduce a new version with a different encoding scheme.  Also, perhaps more importantly, since addresses really just unpack to underlying scripts that enforce a set of constraints, and all software working with an address has to be able to produce scripts that are valid for a given script version, they effectively have a one-to-one correspondence with the underlying script version.

It seems like it would be worthwhile to add a blurb about address versioning in the primer.

> Also, in the PR commit message, Paragraph 2 has a typo:

Fixed.