@davecgh 2021-03-06
@matthawkins90 I believe I have addressed all of you feedback.  See the diff of changes [here](https://github.com/decred/dcrd/compare/06c4018038759914fb3abca6e3b54232949413e4..63de7c801a94c23410a46b8a4cf308b47eb5632e).