@matthawkins90 2021-03-06
Thanks for the updates. Reading now!
Meanwhile, Let me clarify what I meant by

> > You use `script` instead of `Script`.
> 
> Not sure I follow you here. Is there a specific instance you can point out?
> 

Sorry for the confusion. This is incredibly nitpicky and inconsequential, so please let me know if this simply doesn't matter.
Here are my definitions:
1. Capitalized `Script`: the proper noun for the language system
2. Lowercase `script`: a stack of opcodes for a transaction.
3. `txscript`: dcrd's packaged implementation for `Script`.

So, examples:
* Line 20 `- The script version of the aforementioned script`
Should probably read:
`- The Script version of the aforementioned script`
right?

* Line 23: `It is also worth keeping in mind that the script version may be implicit`
Should be `Script`?

* Line 99: `payment scripts that correspond to the same script version the original creator`
Should be `Script version`?

* Line 104
* Line 105
* Line 110
...
I can do these in-line, but I want to make sure I'm not misunderstanding or wasting your time. This was simply something I noticed. Thanks for pointing me to the txscript package and answering my other questions!