@davecgh 2021-03-06
I've updated the generic decoding logic to implement a length check for determining if an address is probably a version 0 base58 address as well as slightly modified it a bit to be a bit more efficient when a new version is added as well as introduced some additional tests to cover the differences.

[See Changes](https://github.com/decred/dcrd/compare/371ed3c11e86c3d1afe15b917a8e5f61961b4d66..43e703410a51d458169b5d501d9ba2c75a87c4eb).