@davecgh 2021-03-07
@matthawkins90 Thanks for the clarification.  I'll think on it a bit as I understand from you description why you find it confusing, though I'm not really sure that `Script` correctly refers to the scripting language because the current incarnation is named `txscript`.  All scripts are accompanied by a script version (aka the scripting language version) that identifies the exact language and semantics the script must use.  Currently, the only supported version is 0, but that is going to change in the future.