@davecgh 2021-03-08
@matthawkins90 After thinking about it a bit, I decided to go with "scripting language version" throughout which should hopefully address your points.  I also changed the wording from `until absolutely necessary` to `until the funds are redeemed` which hopefully resolves that point of confusion.  [Changes here](https://github.com/decred/dcrd/compare/43e703410a51d458169b5d501d9ba2c75a87c4eb..0a83984cad486ec0ce34e78b6d6bd01f01924890).