@jholdstock 2021-03-08
Just reviewed the README. As somebody with a low-medium level of familiarity of the topic, it reads well and goes into a good level of depth. ACK.