@davecgh 2021-03-10
> This probably sounds wierd, but I can't get the example test to print out with `go test --run=ExampleDecodeAddress -v` **unless** I add another `fmt.Println()` somewhere else in the file. Not sure what it is.

It's a testable example.  `go test` compiles, runs, and captures the output and then compares that output against the comments to ensure the example actually produces the output it claims to.  That way they don't go stale as the code changes.