@JoeGruffins 2021-03-10
I see! I don't know how I missed this part of golang testing. I see that the output must match what is under `// Output:` Thank you.